some cleanup and quieting the logs
experiencing an undefined offset error each time a gadget is edited and
saved. this has to do with how the java interacts with the gadget, but
its not an actual issue. for now I have silenced it using @ when loading
the variables from the arrays. also put array definitions in just
because... cleaned up the formatting.